plugins/plugin-tee/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elizaos-plugin-tee"
version = "1.0.0"
description = "elizaOS TEE Plugin - Trusted Execution Environment integration for secure key management and remote attestation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "elizaOS Team", email = "team@elizaos.ai" }
]
keywords = ["ai", "agents", "tee", "attestation", "security", "phala", "elizaos"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "cryptography>=43.0.0",
    "pynacl>=1.5.0",
    "web3>=7.0.0",
    "eth-account>=0.13.0",
    "solders>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.19.0",
    "ruff>=0.14.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/elizaos/eliza"
Documentation = "https://elizaos.ai/docs"
Repository = "https://github.com/elizaos/eliza"

[tool.hatch.build.targets.wheel]
packages = ["elizaos_plugin_tee"]

[tool.hatch.build.targets.sdist]
include = [
    "/elizaos_plugin_tee",
    "/tests",
    "/README.md",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=elizaos_plugin_tee --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
disallow_any_generics = true
disallow_subclassing_any = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "T20", # flake8-print
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "ANN101", # missing type for self
    "ANN102", # missing type for cls
    "S101",   # use of assert (needed for tests)
]

[tool.ruff.lint.isort]
known-first-party = ["elizaos_plugin_tee"]

